Title: Update __str__ method in Re2Shield class
Body:
The __str__ method in the Re2Shield class has been updated to return a string representation that includes the version of the Re2Shield object and the number of patterns in the database. If the database has not been compiled, the number of patterns is displayed as 0. This change ensures that a string is always returned when the Re2Shield object is printed as a string.